requirements.txt
google_api_python_client==2.97.0
google_auth_oauthlib==0.4.6
moviepy==1.0.3
numpy==1.23.0
openai==0.27.8
opencv_python==4.5.5.64
pandas==1.4.3
protobuf==4.24.2
pycryptodome==3.18.0
pytz==2022.1
termcolor==2.3.0
